requirements.txt
altair==4.1.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.5
blinker==1.4
boto3==1.13.1
botocore==1.16.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
Fiona==1.8.13.post1
geopandas==0.3.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.6.0
plotly-geo==1.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pydeck==0.3.1
Pygments==2.6.1
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
pyshp==1.2.10
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.0
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
streamlit==0.58.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
